migrate Settings feature
